выносит общие методы в отдельный класс Smart
